cherry picked merge from chenglong-learn
On chenglongs branch, I wiped the results that he got from merging main
and taking the results (in both experiments.json and the dirs
themselves).

Then I merged those changes in and cherry picked out the deletions (some
but not all were in the main branch) and discarded the experiments.json
as well.

Then finally to get the experiments keys back, I copied them from
chenglong-learn since the set there only contained 2 and 3 client tests,
and appended them onto the end of experiments.json in order to resolve
the merge conflict.

Documenting the process in case I deleted some experiments, we can go
back and try and reverse the process/redo it.